Break recursive type tests into separate file
