+++ test/types/community/collection/findX.test-d.ts
@@ -348,175 +348,3 @@ expectType<WithId<{ a: number; b: string }> | null>(
 expectType<WithId<{ a: number; b: string }> | null>(
   (await coll.findOneAndDelete({ a: 3 }, { projection: { _id: 0 } })).value
 );
-
-/**
- * mutually recursive types are not supported and will not get type safety
- */
-interface A {
-  b: B;
-}
-
-interface B {
-  a: A;
-}
-
-declare const mutuallyRecursive: Collection<A>;
-//@ts-expect-error
-mutuallyRecursive.find({});
-mutuallyRecursive.find({
-  b: {}
-});
-
-/**
- * types that are not recursive in name but are recursive in structure are
- *   still supported
- */
-interface RecursiveButNotReally {
-  a: { a: number; b: string };
-  b: string;
-}
-
-declare const recursiveButNotReallyColl: Collection<RecursiveButNotReally>;
-expectError(
-  recursiveButNotReallyColl.find({
-    'a.a': 'asdf'
-  })
-);
-recursiveButNotReallyColl.find({
-  'a.a': 2
-});
-
-/**
- * recursive schemas are now supported, but with limited type checking support
- */
-interface RecursiveSchema {
-  name: RecursiveSchema;
-  age: number;
-}
-
-declare const collection3: Collection<RecursiveSchema>;
-collection3.find({
-  name: {
-    name: {
-      age: 23
-    }
-  }
-});
-
-collection3.find({
-  age: 23
-});
-
-/**
- * Recursive optional schemas are also supported with the same capabilities as
- *   standard recursive schemas
- */
-interface RecursiveOptionalSchema {
-  name?: RecursiveOptionalSchema;
-  age: number;
-}
-
-declare const collection4: Collection<RecursiveOptionalSchema>;
-
-collection4.find({
-  name: {
-    name: {
-      age: 23
-    }
-  }
-});
-
-collection4.find({
-  age: 23
-});
-
-/**
- * recursive union types are supported
- */
-interface Node {
-  next: Node | null;
-}
-
-declare const nodeCollection: Collection<Node>;
-
-nodeCollection.find({
-  next: null
-});
-
-expectError(
-  nodeCollection.find({
-    next: 'asdf'
-  })
-);
-
-nodeCollection.find({
-  'next.next': 'asdf'
-});
-
-nodeCollection.find({ 'next.next.next': 'yoohoo' });
-
-/**
- * Recursive schemas with arrays are also supported
- */
-interface MongoStrings {
-  projectId: number;
-  branches: Branch[];
-  twoLevelsDeep: {
-    name: string;
-  };
-}
-
-interface Branch {
-  id: number;
-  name: string;
-  title?: string;
-  directories: Directory[];
-}
-
-interface Directory {
-  id: number;
-  name: string;
-  title?: string;
-  branchId: number;
-  files: (number | Directory)[];
-}
-
-declare const coll5: Collection<MongoStrings>;
-expectError(
-  coll5.findOne({
-    'branches.0.id': 'hello'
-  })
-);
-
-expectError(
-  coll5.findOne({
-    'branches.0.directories.0.id': 'hello'
-  })
-);
-
-// type safety breaks after the first
-//   level of nested types
-coll5.findOne({
-  'branches.0.directories.0.files.0.id': 'hello'
-});
-
-coll5.findOne({
-  branches: [
-    {
-      id: 'asdf'
-    }
-  ]
-});
-
-// type inference works on properties but only at the top level
-expectError(
-  coll5.findOne({
-    projectId: 'asdf'
-  })
-);
-
-coll5.findOne({
-  twoLevelsDeep: {
-    name: 3
-  }
-});
